goal receptacle segmentation for place, waypoint 2 velocity controller update
